// client/src/types/ActionTypes.ts
export enum AT {
    START_SCORE_ASYNC = 'START_SCORE_ASYNC',
    START_SCORE_FULFILLED = 'START_SCORE_FULFILLED',
    START_SCORE_REJECTED = 'START_SCORE_REJECTED',

    ADD_SCORE_ASYNC = 'ADD_SCORE_ASYNC',
    ADD_SCORE_FULFILLED = 'ADD_SCORE_FULFILLED',
    ADD_SCORE_REJECTED = 'ADD_SCORE_REJECTED',
    
    USER_LOGIN = 'USER_LOGIN',
    USER_LOGOUT = 'USER_LOGOUT',

    GET_USERS_ASYNC = 'GET_USERS_ASYNC',
    GET_USERS_FULFILLED = 'GET_USERS_FULFILLED',
    GET_USERS_REJECTED = 'GET_USERS_REJECTED',

    GET_SCORES_ASYNC = 'GET_SCORES_ASYNC',
    GET_SCORES_FULFILLED = 'GET_SCORES_FULFILLED',
    GET_SCORES_REJECTED = 'GET_SCORES_REJECTED',

    GET_SCORE_DETAILS_ASYNC = 'GET_SCORE_DETAILS_ASYNC',
    GET_SCORE_DETAILS_FULFILLED = 'GET_SCORE_DETAILS_FULFILLED',
    GET_SCORE_DETAILS_REJECTED = 'GET_SCORE_DETAILS_REJECTED',

    GET_USER_SCORES_ASYNC = 'GET_USER_SCORES_ASYNC',
    GET_USER_SCORES_FULFILLED = 'GET_USER_SCORES_FULFILLED',
    GET_USER_SCORES_REJECTED = 'GET_USER_SCORES_REJECTED',

    GET_TABLE_TYPES_ASYNC = 'GET_TABLE_TYPES_ASYNC',
    GET_TABLE_TYPES_FULFILLED = 'GET_TABLE_TYPES_FULFILLED',
    GET_TABLE_TYPES_REJECTED = 'GET_TABLE_TYPES_REJECTED',
}

export default AT